Module Timere.Time_zone

type t

val make : string -> (t, unit) Stdlib.result

Makes a time zone from name.

Naming follows the convention used in /usr/share/zoneinfo/posix/ distributed on Linux, e.g. "Australia/Sydney".

See available_time_zones or checking usable time zone names at runtime.

Alternatively, if you are using timere.tz.full (the default implementation for timere.tz.data), then you can also see available-time-zones.txt for available time zones.

val make_exn : string -> t

raises Invalid_argument

if make fails

val name : t -> string

val utc : t

val equal : t -> t -> bool

val available_time_zones : string list

val make_offset_only : ?⁠name:string -> int -> t

This is mainly used for when you only have an offset to work with, and you don't need to do any accurate search over time zones.

One use of this is to create a time zone for to_string functions.

Importing and exporting

type entry = {

is_dst : bool;

offset : int;

}

val of_transitions : name:string -> (int64 * entry) list -> (t, unit) Stdlib.result

val transitions : t -> ((int64 * int64) * entry) list

val transition_seq : t -> ((int64 * int64) * entry) Stdlib.Seq.t

val to_sexp : t -> CCSexp.t

val of_sexp : CCSexp.t -> (t, unit) Stdlib.result

val to_sexp_string : t -> string

val of_sexp_string : string -> (t, unit) Stdlib.result

val to_json_string : t -> string

val of_json_string : string -> (t, unit) Stdlib.result