Module Timere.Timestamp

val min : timestamp

val max : timestamp

val now : unit -> timestamp

val pp : ?⁠display_using_tz:Time_zone.t -> ?⁠format:string -> unit -> Stdlib.Format.formatter -> timestamp -> unit

Pretty printing for timestamp.

Follows same format string rules and default format string as Date_time.to_string.

val to_string : ?⁠display_using_tz:Time_zone.t -> ?⁠format:string -> timestamp -> string

val pp_rfc3339 : ?⁠frac_s:int -> unit -> Stdlib.Format.formatter -> timestamp -> unit

frac_s determines the number of fractional digits to include.

raises Invalid_argument

if frac_s < 0 || frac_s > 9

val pp_rfc3339_milli : Stdlib.Format.formatter -> timestamp -> unit

val pp_rfc3339_micro : Stdlib.Format.formatter -> timestamp -> unit

val pp_rfc3339_nano : Stdlib.Format.formatter -> timestamp -> unit

val to_rfc3339 : ?⁠frac_s:int -> timestamp -> string

val to_rfc3339_milli : timestamp -> string

val to_rfc3339_micro : timestamp -> string

val to_rfc3339_nano : timestamp -> string

val of_iso8601 : string -> (timestamp, string) Stdlib.result