Module Timere

type t

This is the core type of Timere used to encode computation over time.

The following documentation may call value of type timere "a Timere object".

exception Invalid_format_string of string

Printing exception

type tz_offset_s = int

type timestamp = int64

type 'a range = [

| `Range_inc of 'a * 'a

| `Range_exc of 'a * 'a

]

Basic constructors

val always : t

Entire interval that Timere can handle, i.e. [min_timestamp, max_timestamp\)

val empty : t

Empty interval

val years : int list -> t

years l is a shorthand for pattern ~years:l ()

type month = [

| `Jan

| `Feb

| `Mar

| `Apr

| `May

| `Jun

| `Jul

| `Aug

| `Sep

| `Oct

| `Nov

| `Dec

]

val months : month list -> t

months l is a shorthand for pattern ~months:l ()

val month_days : int list -> t

month_days l is a shorthand for pattern ~month_days:l ()

type weekday = [

| `Sun

| `Mon

| `Tue

| `Wed

| `Thu

| `Fri

| `Sat

]

val weekdays : weekday list -> t

weekdays l is a shorthand for pattern ~weekdays:l ()

val hours : int list -> t

hours l is a shorthand for pattern ~hours:l ()

val minutes : int list -> t

minutes l is a shorthand for pattern ~minutes:l ()

val seconds : int list -> t

seconds l is a shorthand for pattern ~seconds:l ()

val pattern : ?⁠strict:bool -> ?⁠years:int list -> ?⁠year_ranges:int range list -> ?⁠months:month list -> ?⁠month_ranges:month range list -> ?⁠month_days:int list -> ?⁠month_day_ranges:int range list -> ?⁠weekdays:weekday list -> ?⁠weekday_ranges:weekday range list -> ?⁠hours:int list -> ?⁠hour_ranges:int range list -> ?⁠minutes:int list -> ?⁠minute_ranges:int range list -> ?⁠seconds:int list -> ?⁠second_ranges:int range list -> unit -> t

Pattern matches over date times.

A pattern p matches date time dt if

(dt.year is in p.years or p.year_ranges)
&& (dt.month is in p.months or p.month_ranges)
&& (dt.month_day is in p.month_days or p.month_day_ranges)
&& (dt.weekday is in p.weekdays or p.weekday_ranges)
&& (dt.hour is in p.hours or p.hour_ranges)
&& (dt.minute is in p.minutes or p.minute_ranges)
&& (dt.second is in p.seconds or p.second_ranges)

Empty pattern levels are treated as wildcard, e.g. if p.years and p.year_ranges are both empty, then (dt.year is in p.years or p.year_ranges) is true.

Time zone

module Time_zone : sig ... end

val with_tz : Time_zone.t -> t -> t

with_tz tz t changes the time zone to evaluate t in to tz

Algebraic operations

val inter : t list -> t

val union : t list -> t

val not : t -> t

Duration

module Duration : sig ... end

val shift : Duration.t -> t -> t

val lengthen : Duration.t -> t -> t

Discrete time points

val cur_timestamp : unit -> int64

val min_timestamp : int64

val max_timestamp : int64

exception Invalid_timestamp

val of_timestamps : ?⁠skip_invalid:bool -> timestamp list -> t

val of_timestamp_seq : ?⁠skip_invalid:bool -> timestamp Stdlib.Seq.t -> t

module Date_time : sig ... end

List and Filtering operations

type chunked

type chunking = [

| `Disjoint_intervals

| `By_duration of Duration.t

| `By_duration_drop_partial of Duration.t

| `At_year_boundary

| `At_month_boundary

]

val chunk : chunking -> (chunked -> chunked) -> t -> t

val chunk_again : chunking -> chunked -> chunked

val first : chunked -> chunked

val take : int -> chunked -> chunked

val take_nth : int -> chunked -> chunked

val drop : int -> chunked -> chunked

val first_point : t -> t

val take_n_points : int -> t -> t

val drop_n_points : int -> t -> t

Manual intervals

val interval_dt_inc : Date_time.t -> Date_time.t -> t

val interval_dt_exc : Date_time.t -> Date_time.t -> t

val interval_inc : timestamp -> timestamp -> t

val interval_exc : timestamp -> timestamp -> t

exception Interval_is_invalid

exception Interval_is_empty

exception Intervals_are_not_sorted

exception Intervals_are_not_disjoint

type interval = timestamp * timestamp

val of_intervals : ?⁠skip_invalid:bool -> interval list -> t

val of_interval_seq : ?⁠skip_invalid:bool -> interval Stdlib.Seq.t -> t

val of_sorted_intervals : ?⁠skip_invalid:bool -> interval list -> t

val of_sorted_interval_seq : ?⁠skip_invalid:bool -> interval Stdlib.Seq.t -> t

Hour minute second intervals

type hms = private {

hour : int;

minute : int;

second : int;

}

val make_hms : hour:int -> minute:int -> second:int -> (hms, unit) Stdlib.result

val make_hms_exn : hour:int -> minute:int -> second:int -> hms

val hms_interval_inc : hms -> hms -> t

val hms_interval_exc : hms -> hms -> t

val of_hms_intervals : (hms * hms) Stdlib.Seq.t -> t

Search oriented operations

val after : Duration.t -> t -> t -> t

val between_inc : Duration.t -> t -> t -> t

val between_exc : Duration.t -> t -> t -> t

Infix operators

module Infix : sig ... end

Resolution

val resolve : ?⁠search_using_tz:Time_zone.t -> t -> (interval Stdlib.Seq.t, string) Stdlib.result

Resolves a Timere object into a concrete interval sequence

Pretty printers

val sprintf_timestamp : ?⁠display_using_tz:Time_zone.t -> ?⁠format:string -> timestamp -> string

val pp_timestamp : ?⁠display_using_tz:Time_zone.t -> ?⁠format:string -> Stdlib.Format.formatter -> timestamp -> unit

val sprintf_interval : ?⁠display_using_tz:Time_zone.t -> ?⁠format:string -> interval -> string

val pp_interval : ?⁠display_using_tz:Time_zone.t -> ?⁠format:string -> Stdlib.Format.formatter -> interval -> unit

S-expressions

val to_sexp : t -> CCSexp.t

val pp_sexp : Stdlib.Format.formatter -> t -> unit

val to_sexp_string : t -> string

val of_sexp : CCSexp.t -> (t, string) Stdlib.result

val of_sexp_string : string -> (t, string) Stdlib.result

Misc

module Utils : sig ... end