Module Timere

Introduction

It is advised to read through this to understand why certain API seem needlessly complicated compared to alternatives (possibly in another programming language).

Misconceptions

Time zone offsets are always in hours

What we typically consider a time zone, e.g. "Europe/Paris", always has a constant offset

With a time zone and a specific date time, we can obtain a unique "unix timestamp" (time since unix epoch)

We can calculate time zone offset at some date time, and apply it universally for any other date time in the same time zone, and this always works

Many more on various online resources...

Time zone, time zone offset, and date time

It is tempting to think that a time zone maps cleanly to a constant offset, and indeed we may define time zone as such, e.g. UTC+1, UTC-10, but this is far from what we mean in everyday context.

Very often, what we consider to be time zone actually represents a table which records what offset to use in which period, which we index/refer to by geographical names like "Europe/Paris", "Australia/Sydney". These tables are defined by governmental bodies, and attributes of the table, such as offset of any particular period, start and end of any particular period, may not show any observable pattern.

Thus it is not uncommon to see date time errors arising from attempts of applying some formulas universally, which might work well for a lot of cases in contemporary time periods, but fail for some combinations.

We make explicit of above explanation by considering "Europe/Paris" as an example, which observes a common form of transition called Daylight Saving Time (DST).

When DST starts (usually in March), the clocks "jump forward" by 1 hour, usually jumping from 2am to 3am, leading 2am to 3am (exclusive) to become non-existent.

Indeed we can observe the lack of continuity of Europe/Paris timeline below (UTC timeline is always continuous):

                         Mar
UTC          -------------|-------------
                         1am

Europe/Paris -------------|-------------
                       2am 3am
                      (+1) (+2)

Paris time zone offset also changes from UTC+1 to UTC+2.

When DST ends (usually in Oct), clocks "jump backward" by 1 hour, usually jumping from 3am to 2am, leading to 2am to 3am (exclusive) becoming duplicated:

                         Oct
UTC          -------------|-------------
                         1am

Europe/Paris -------------|-------------
                       3am 2am
                      (+2) (+1)

Paris time zone offset also changes from UTC+2 to UTC+1.

Another way of looking at above is when DST is in effect, Paris observes UTC+2, and UTC+1 otherwise:

                          |-------------DST on------------|
             |---DST off--|                               |---DST off--|

                         Mar                             Oct
UTC          -------------|------------- ... -------------|-------------
                         1am                             1am

Europe/Paris -------------|------------- ... -------------|-------------
                       2am 3am                         3am 2am
                      (+1) (+2)                       (+2) (+1)

This start and end of the DST on and off periods, along with the corresponding offsets, form the basis of the table we mentioned above.

Timere date time API basics

We highlight some critical cases in practice, and how Timere behaves and how it may differ from other libraries.

Take year 2021 for example, DST starts on 2021 Mar 28 for Paris, causing clocks to jump from 2am to 3am. Pick any intermediate point, say 2:30am, we yield an undefined date time. In this case, Timere refuses the construction of such Date_time.t in Date_time.make etc, while some libraries coerce the result into 3:30am.

And DST ends on 2021 Oct 31, causing clocks to jump from 3am to 2am. Say we pick 2:30am again, we are actually pointing at two time points (there are two 2:30am) unless we make an explicit selection between the first or second occurance. Whenever ambiguity of this form is a possiblity for the result of a function, say Date_time.to_timestamp, Timere uses Date_time.local_result variant type, of which `Single _ indicates lack of ambiguity for the particular result, and `Ambiguous _ indicates the result is ambiguous.

Some other libraries coerce the ambiguous result into one of the two possible choices (which exact one may not be guaranteed). If user wishes to do similar coercions, they may use Date_time.min_of_local_result or Date_time.max_of_local_result.

For constructions, Date_time.make yields a possibly ambiguous construction, while Date_time.make_unambiguous yields an unambiguous construction. In general, if you are provided with the exact offset to UTC, then make_unambiguous is the better choice.

Span

module Span : sig ... end

Date time handling

type weekday = [

| `Sun

| `Mon

| `Tue

| `Wed

| `Thu

| `Fri

| `Sat

]

exception Invalid_format_string of string

Printing exception

type timestamp = Span.t

module Time_zone : sig ... end

module Date_time : sig ... end

module Timestamp : sig ... end

module Week_date_time : sig ... end

Reasoning/scheduling

type t

This is the core type of Timere that represents sets of points in time, more precisely, unions of time intervals. For example, "all Mondays of year 2000 at the UTC timezone".

We call Timere.t values "timere object"; internally they are rich expressions representing the time computations (union, intersection, etc.), lazily forced into more low-level descriptions (lazy sequences of intervals).

type 'a range = [

| `Range_inc of 'a * 'a

| `Range_exc of 'a * 'a

]

Basic constructors

val now : unit -> t

Time right now

val always : t

Entire interval that Timere can handle, i.e. [0000 Jan 01 14:00:00 +00:00:00, 9999 Dec 31 09:59:58 +00:00:00)

val empty : t

Empty interval

val pattern : ?⁠years:int list -> ?⁠year_ranges:int range list -> ?⁠months:int list -> ?⁠month_ranges:int range list -> ?⁠days:int list -> ?⁠day_ranges:int range list -> ?⁠weekdays:weekday list -> ?⁠weekday_ranges:weekday range list -> ?⁠hours:int list -> ?⁠hour_ranges:int range list -> ?⁠minutes:int list -> ?⁠minute_ranges:int range list -> ?⁠seconds:int list -> ?⁠second_ranges:int range list -> unit -> t

Pattern matches over date times.

A pattern p matches date time dt if

(dt.year is in p.years or p.year_ranges)
&& (dt.month is in p.months or p.month_ranges)
&& (dt.month_day is in p.month_days or p.month_day_ranges)
&& (dt.weekday is in p.weekdays or p.weekday_ranges)
&& (dt.hour is in p.hours or p.hour_ranges)
&& (dt.minute is in p.minutes or p.minute_ranges)
&& (dt.second is in p.seconds or p.second_ranges)

Empty pattern levels are treated as wildcard, e.g. if p.years and p.year_ranges are both empty, then (dt.year is in p.years or p.year_ranges) is true.

val years : int list -> t

years l is a shorthand for pattern ~years:l ()

val year_ranges : int range list -> t

year_ranges l is a shorthand for pattern ~year_ranges:l ()

val months : int list -> t

months l is a shorthand for pattern ~months:l ()

val month_ranges : int range list -> t

month_ranges l is a shorthand for pattern ~month_ranges:l ()

val days : int list -> t

days l is a shorthand for pattern ~month_days:l ()

val day_ranges : int range list -> t

day_ranges l is a shorthand for pattern ~month_day_ranges:l ()

val weekdays : weekday list -> t

weekdays l is a shorthand for pattern ~weekdays:l ()

val weekday_ranges : weekday range list -> t

weekday_ranges l is a shorthand for pattern ~weekday_ranges:l ()

val hours : int list -> t

hours l is a shorthand for pattern ~hours:l ()

val hour_ranges : int range list -> t

hour_ranges l is a shorthand for pattern ~hour_ranges:l ()

val minutes : int list -> t

minutes l is a shorthand for pattern ~minutes:l ()

val minute_ranges : int range list -> t

minute_ranges l is a shorthand for pattern ~minute_ranges:l ()

val seconds : int list -> t

seconds l is a shorthand for pattern ~seconds:l ()

val second_ranges : int range list -> t

second_ranges l is a shorthand for pattern ~second_ranges:l ()

val nth_weekday_of_month : int -> weekday -> t

nth_weekday_of_month n wday picks the nth weekday of all months, where 1 <= n && n <= 5

raises Invalid_argument

if n is out of range

Algebraic operations

val inter : t list -> t

Intersection of list of timeres.

inter [] is equivalent to always.

val union : t list -> t

Union of list of timeres.

union [] is equivalent to empty.

val not : t -> t

Negation of timere.

not t is equivalent to all the intervals not included in t.

val shift : Span.t -> t -> t

val lengthen : Span.t -> t -> t

raises Invalid_argument

if duration is negative

val with_tz : Time_zone.t -> t -> t

with_tz tz t changes the time zone to evaluate t in to tz

val date_time : Date_time.t -> t

val before : Date_time.t -> t

val since : Date_time.t -> t

val after : Date_time.t -> t

val date_times : Date_time.t list -> t

val date_time_seq : Date_time.t Stdlib.Seq.t -> t

val sorted_date_times : Date_time.t list -> t

val sorted_date_time_seq : Date_time.t Stdlib.Seq.t -> t

exception Invalid_timestamp

val timestamp : timestamp -> t

val before_timestamp : timestamp -> t

val since_timestamp : timestamp -> t

val after_timestamp : timestamp -> t

val timestamps : ?⁠skip_invalid:bool -> timestamp list -> t

timestamps l

skip_invalid defaults to false

raises Invalid_timestamp

if not skip_invalid and l contains an invalid timestamp

val timestamp_seq : ?⁠skip_invalid:bool -> timestamp Stdlib.Seq.t -> t

timestamps s

skip_invalid defaults to false

raises Invalid_timestamp

if not skip_invalid and s contains an invalid timestamp

val sorted_timestamps : ?⁠skip_invalid:bool -> timestamp list -> t

val sorted_timestamp_seq : ?⁠skip_invalid:bool -> timestamp Stdlib.Seq.t -> t

Intervals

Explicit

exception Interval_is_invalid

exception Intervals_are_not_sorted

module Interval : sig ... end

val intervals : ?⁠skip_invalid:bool -> Interval.t list -> t

intervals l

skip_invalid defaults to false

raises Interval_is_invalid

if not skip_invalid and l contains an invalid interval

val interval_seq : ?⁠skip_invalid:bool -> Interval.t Stdlib.Seq.t -> t

interval_seq s

skip_invalid defaults to false

raises Interval_is_invalid

if not skip_invalid and s contains an invalid interval

val sorted_intervals : ?⁠skip_invalid:bool -> Interval.t list -> t

sorted_intervals l

skip_invalid defaults to false

raises Interval_is_invalid

if not skip_invalid and l contains an invalid interval

raises Intervals_are_not_sorted

if l is not sorted

val sorted_interval_seq : ?⁠skip_invalid:bool -> Interval.t Stdlib.Seq.t -> t

sorted_interval_seq s

skip_invalid defaults to false

raises Interval_is_invalid

if not skip_invalid and s contains an invalid interval

raises Intervals_are_not_sorted

if s is not sorted

Pattern matching

module Points : sig ... end

type points = Points.t

val bounded_intervals : [ `Whole | `Snd ] -> Span.t -> points -> points -> t

bounded_intervals mode bound p1 p2 for each point x matched by p1, then for the earliest point y matched by p2 such that x < y && y - x <= bound

if mode = `Whole, yields (x, y)

if mode = `Snd, yields (y, y + 1)

Examples:

bounded_intervals `Whole (Span.For_human.make ~days:1 ())
  (make_points ~hour:13 ~minute:0 ~second:0 ()) (* p1 *)
  (make_points ~hour:14 ~minute:0 ~second:0 ()) (* p2 *)

yields all the "1pm to 2pm" intervals, since at each "1pm" mark represented by p1, searching forward up to 24 hour period, we can find a "2pm" mark in p2

bounded_intervals `Whole (Span.For_human.make ~days:1 ())
  (make_points ~month:2 ~day:10 ~hour:13 ~minute:0 ~second:0 ()) (* p1 *)
  (make_points                  ~hour:14 ~minute:0 ~second:0 ()) (* p2 *)

yields all the "Feb 10th 1pm to 2pm" intervals (or specifically "Feb 10th 1pm to Feb 10th 2pm")

bounded_intervals `Whole (Span.For_human.make ~days:1 ())
  (make_points ~month:`Feb ~day:10 ~hour:23 ~minute:0 ~second:0 ()) (* p1 *)
  (make_points                     ~hour:3  ~minute:0 ~second:0 ()) (* p2 *)

yields all the "Feb 10th 11pm to 3am" intervals (or specifically "Feb 10th 11pm to Feb 11th 3am")

raises Invalid_argument

if bound is negative

raises Invalid_argument

if precision (number of date time arguments passed to make_points during construction) of p1 < precision of p2

For example, make_points_exn ~hour:3 ~minute:0 ~second:0 () has a lower precision than make_points_exn ~day:10 ~hour:12 ~minute:30 ~second:0 ()

Hour minute second intervals

module Hms : sig ... end

val hms_intervals_inc : Hms.t -> Hms.t -> t

Same as hms_intervals_exc ... with end point increased by one second

val hms_intervals_exc : Hms.t -> Hms.t -> t

Same as bounded_intervals ... with bound fixed to Span.For_human.make ~days:1 ()

Chunking

type chunked

type chunking = [

| `Disjoint_intervals

| `By_duration of Span.t

| `By_duration_drop_partial of Span.t

| `At_year_boundary

| `At_month_boundary

]

Ways to chunk/slice time intervals for the selector.

`Disjoint_intervals gives a sequence of disjoint intervals to the selector, specifically they are in ascending order, non-overlapping, non-connecting, and unique

`By_duration slices in the fixed size specified by the duration. Partial chunks (chunks less than the fixed size) are preserved.

`By_duration_drop_partial slices in the fixed size specified by the duration. Partial chunks (chunks less than the fixed size) are discarded.

`At_year_boundary slices at the year boundary (e.g. 2021 Jan 1st 00:00:00)

`At_month_boundary slices at the month boundary (e.g. Aug 1st 00:00:00)

val chunk : chunking -> (chunked -> chunked) -> t -> t

chunk chunking f t applies chunked selector f on t.

raises Invalid_argument

if duration is negative in `By_duration or `By_duration_drop_partial

Chunked selectors

val chunk_again : chunking -> chunked -> chunked

chunk_again chunking f applies chunked selector f as a selector

val first : chunked -> chunked

Takes only first chunk

val take : int -> chunked -> chunked

Takes n chunks

val take_nth : int -> chunked -> chunked

Take every nth chunk, specifically 0th, nth, 2nth, 3nth, ...

val drop : int -> chunked -> chunked

Discard n chunks

Infix operators

val (&) : t -> t -> t

inter

val (|||) : t -> t -> t

union

val (%>) : ('a -> 'b) -> ('b -> 'c) -> 'a -> 'c

Composition, mainly for chunked selectors

f1 %> f2 is equivalent to fun x -> x |> f1 |> f2.

Resolution

val resolve : ?⁠search_using_tz:Time_zone.t -> t -> (Interval.t Stdlib.Seq.t, string) Stdlib.result

Resolves a Timere object into a concrete interval sequence

Pretty printers

val pp_intervals : ?⁠display_using_tz:Time_zone.t -> ?⁠format:string -> ?⁠sep:unit Fmt.t -> unit -> Stdlib.Format.formatter -> Interval.t Stdlib.Seq.t -> unit

S-expressions

val to_sexp : t -> CCSexp.t

val to_sexp_string : t -> string

val of_sexp : CCSexp.t -> (t, string) Stdlib.result

val of_sexp_string : string -> (t, string) Stdlib.result

val pp_sexp : Stdlib.Format.formatter -> t -> unit

Misc

module Utils : sig ... end