Module Timere.Time_zone

type t

val make : string -> t option

Makes a time zone from name.

Naming follows the convention used in /usr/share/zoneinfo/posix/ distributed on Linux, e.g. "Australia/Sydney".

See available_time_zones or checking usable time zone names at runtime.

Alternatively, if you are using timere.tz.full (the default implementation for timere.tz.data), then you can also see available-time-zones.txt for available time zones.

val make_exn : string -> t

raises Invalid_argument

if make fails

val name : t -> string

val utc : t

val equal : t -> t -> bool

val available_time_zones : string list

val make_offset_only : ?⁠name:string -> int -> t

This is mainly used for when you only have an offset to work with, and you don't need to do any accurate search over time zones.

One use of this is to create a time zone for to_string functions.

type entry = {

is_dst : bool;

offset : int;

}

Importing and exporting

type table = (int64, Stdlib.Bigarray.int64_elt, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t * entry array

module Raw : sig ... end

module Sexp : sig ... end

module JSON : sig ... end

module Db : sig ... end